moving vis/visualize to visualizations plugin
