+++ src/legacy/core_plugins/kbn_vislib_vis_types/public/goal.js
@@ -24,7 +24,7 @@ import { ColorSchemas } from 'ui/vislib/components/color/colormaps';
 import { GaugeOptions } from './components/options';
 import { getGaugeCollections, GaugeTypes, ColorModes } from './utils/collections';
 import { vislibVisController } from './controller';
-import { visFactory } from '../../../ui/public/vis/vis_factory';
+import { visFactory } from '../../visualizations/public/np_ready/public/vis_factory';
 
 export default function GoalVisType() {
