+++ src/legacy/core_plugins/kibana/public/discover/kibana_services.ts
@@ -81,7 +81,7 @@ export function getServices() {
 
 // EXPORT legacy static dependencies
 export { angular };
-export { buildVislibDimensions } from 'ui/visualize/loader/pipeline_helpers/build_pipeline';
+export { buildVislibDimensions } from 'ui/../../../visualizations/public/np_ready/public/legacy/build_pipeline';
 // @ts-ignore
 export { callAfterBindingsWorkaround } from 'ui/compat';
 export {
