+++ src/legacy/core_plugins/vis_type_metric/public/types.ts
@@ -19,7 +19,7 @@
 
 import { ColorSchemas } from 'ui/vislib/components/color/colormaps';
 import { RangeValues } from 'ui/vis/editors/default/controls/ranges';
-import { SchemaConfig } from 'ui/visualize/loader/pipeline_helpers/build_pipeline';
+import { SchemaConfig } from 'ui/../../visualizations/public/np_ready/public/legacy/build_pipeline';
 import { ColorModes } from '../../kbn_vislib_vis_types/public/utils/collections';
 import { Labels, Style } from '../../kbn_vislib_vis_types/public/types';
