+++ src/legacy/core_plugins/vis_type_tagcloud/public/tag_cloud_type.ts
@@ -18,7 +18,7 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { Status } from 'ui/vis/update_status';
+import { Status } from 'ui/../../visualizations/public/np_ready/public/legacy/update_status';
 // @ts-ignore
 import { Schemas } from 'ui/vis/editors/default/schemas';
