+++ src/legacy/core_plugins/visualizations/public/expressions/visualization_renderer.tsx
@@ -22,7 +22,7 @@ import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
 // @ts-ignore
 import { Vis } from '../../../../ui/public/visualize/loader/vis';
-import { Visualization } from '../../../../ui/public/visualize/components';
+import { Visualization } from '../../../visualizations/public/np_ready/public/components';
 
 export const visualization = () => ({
   name: 'visualization',
