+++ src/legacy/core_plugins/visualizations/public/np_ready/public/components/_index.scss
@@ -1,0 +1,1 @@
+@import 'visualization';
