+++ src/legacy/core_plugins/visualizations/public/np_ready/public/components/_visualization.scss
