+++ src/legacy/core_plugins/visualizations/public/np_ready/public/components/visualization_chart.tsx
@@ -21,10 +21,10 @@ import React from 'react';
 import * as Rx from 'rxjs';
 import { debounceTime, filter, share, switchMap } from 'rxjs/operators';
 
-import { PersistedState } from '../../persisted_state';
-import { ResizeChecker } from '../../../../../plugins/kibana_utils/public';
-import { Vis, VisualizationController } from '../../vis';
-import { getUpdateStatus } from '../../vis/update_status';
+import { PersistedState } from '../../../../../../ui/public/persisted_state';
+import { ResizeChecker } from '../../../../../../ui/public/resize_checker';
+import { Vis, VisualizationController } from '../../../../../../ui/public/vis';
+import { getUpdateStatus } from '../legacy/update_status';
 
 interface VisualizationChartProps {
   onInit?: () => void;
