+++ src/legacy/core_plugins/visualizations/public/np_ready/public/components/visualization_noresults.test.js
