+++ src/legacy/core_plugins/visualizations/public/np_ready/public/components/visualization_noresults.tsx
@@ -19,7 +19,7 @@
 
 import { EuiIcon, EuiSpacer, EuiText } from '@elastic/eui';
 import React from 'react';
-import { dispatchRenderComplete } from '../../../../../plugins/kibana_utils/public';
+import { dispatchRenderComplete } from '../../../../../../../plugins/kibana_utils/public';
 
 interface VisualizationNoResultsProps {
   onInit?: () => void;
