+++ src/legacy/core_plugins/visualizations/public/np_ready/public/filters/brush_event.test.js
