+++ src/legacy/core_plugins/visualizations/public/np_ready/public/legacy/calculate_object_hash.d.ts
