+++ src/legacy/core_plugins/visualizations/public/np_ready/public/legacy/vis_update.js
