+++ src/legacy/core_plugins/visualizations/public/np_ready/public/types/base_vis_type.js
