+++ src/legacy/ui/public/vis/vis_types/index.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { BaseVisType } from './base_vis_type';
-import { ReactVisType } from './react_vis_type';
+import { BaseVisType } from '../../../../core_plugins/visualizations/public/np_ready/public/types/base_vis_type';
+import { ReactVisType } from '../../../../core_plugins/visualizations/public/np_ready/public/types/react_vis_type';
 
 export { BaseVisType, ReactVisType };
