+++ src/legacy/ui/public/visualize/_index.scss
@@ -1,1 +1,1 @@
-@import './components/index';
+@import '../../../core_plugins/visualizations/public/np_ready/public/components/index';
