--- src/legacy/ui/public/visualize/components/_index.scss
@@ -1,1 +1,0 @@
-@import './visualization';
