+++ src/legacy/ui/public/visualize/loader/pipeline_helpers/index.ts
@@ -17,4 +17,4 @@
  * under the License.
  */
 
-export { buildPipeline } from './build_pipeline';
+export { buildPipeline } from '../../../../../core_plugins/visualizations/public/np_ready/public/legacy/build_pipeline';
